+++ flink-formats/flink-orc/src/test/java/org/apache/flink/orc/OrcRowInputFormatTest.java
@@ -26,7 +26,6 @@ import org.apache.flink.api.java.typeutils.ObjectArrayTypeInfo;
 import org.apache.flink.api.java.typeutils.RowTypeInfo;
 import org.apache.flink.core.fs.FileInputSplit;
 import org.apache.flink.core.fs.Path;
-import org.apache.flink.orc.util.OrcTestFileGenerator;
 import org.apache.flink.types.Row;
 import org.apache.flink.util.InstantiationUtil;
 
@@ -35,7 +34,9 @@ import org.apache.hadoop.hive.ql.io.sarg.PredicateLeaf;
 import org.apache.hadoop.hive.ql.io.sarg.SearchArgument;
 import org.apache.orc.Reader;
 import org.apache.orc.StripeInformation;
+import org.apache.orc.TypeDescription;
 import org.junit.After;
+import org.junit.Assert;
 import org.junit.Test;
 
 import java.io.FileNotFoundException;
@@ -46,6 +47,7 @@ import java.sql.Timestamp;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
+import java.util.stream.IntStream;
 
 import static org.junit.Assert.assertArrayEquals;
 import static org.junit.Assert.assertEquals;
@@ -53,12 +55,8 @@ import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
-import static org.mockito.Matchers.eq;
-import static org.mockito.Mockito.any;
-import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.spy;
-import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 /**
@@ -126,7 +124,7 @@ public class OrcRowInputFormatTest {
 	private static final String TEST_FILE_NESTEDLIST = "test-data-nestedlist.orc";
 	private static final String TEST_SCHEMA_NESTEDLIST = "struct<mylist1:array<array<struct<mylong1:bigint>>>>";
 
-	/** Generated by {@link OrcTestFileGenerator#writeCompositeTypesWithNullsFile(String)}. */
+	/** Generated by {@code OrcTestFileGenerator#writeCompositeTypesWithNullsFile(String)}. */
 	private static final String TEST_FILE_COMPOSITES_NULLS = "test-data-composites-with-nulls.orc";
 	private static final String TEST_SCHEMA_COMPOSITES_NULLS =
 		"struct<" +
@@ -136,7 +134,7 @@ public class OrcRowInputFormatTest {
 			"list2:array<map<string,int>>" +
 		">";
 
-	/** Generated by {@link OrcTestFileGenerator#writeCompositeTypesWithRepeatingFile(String)}. */
+	/** Generated by {@code OrcTestFileGenerator#writeCompositeTypesWithRepeatingFile(String)}. */
 	private static final String TEST_FILE_REPEATING = "test-data-repeating.orc";
 	private static final String TEST_SCHEMA_REPEATING =
 		"struct<" +
@@ -184,10 +182,6 @@ public class OrcRowInputFormatTest {
 
 		OrcRowInputFormat spy = spy(rowOrcInputFormat);
 
-		// mock options to check configuration of ORC reader
-		Reader.Options options = new Reader.Options();
-		doReturn(options).when(spy).getOptions(any());
-
 		spy.selectFields(9, 11, 2);
 		spy.openInputFormat();
 		FileInputSplit[] splits = spy.createInputSplits(1);
@@ -202,7 +196,7 @@ public class OrcRowInputFormatTest {
 			true, true, true, true, true, // nested field 9 is in
 			false, false, false, false, // nested field 10 is out
 			true, true, true, true, true}; // nested field 11 is in
-		assertArrayEquals(expected, options.getInclude());
+		assertArrayEquals(expected, spy.getReader().getOptions().getInclude());
 	}
 
 	@Test
@@ -212,28 +206,45 @@ public class OrcRowInputFormatTest {
 
 		OrcRowInputFormat spy = spy(rowOrcInputFormat);
 
-		// mock options to check configuration of ORC reader
-		Reader.Options options = spy(new Reader.Options());
-		doReturn(options).when(spy).getOptions(any());
-
 		FileInputSplit[] splits = spy.createInputSplits(3);
 
 		spy.openInputFormat();
 		spy.open(splits[0]);
-		verify(options).range(eq(3L), eq(137005L));
+		assertOptions(spy.getReader(), 3L, 137005L);
 		spy.open(splits[1]);
-		verify(options).range(eq(137008L), eq(136182L));
+		assertOptions(spy.getReader(), 137008L, 136182L);
 		spy.open(splits[2]);
-		verify(options).range(eq(273190L), eq(123633L));
+		assertOptions(spy.getReader(), 273190L, 123633L);
+	}
+
+	private static void assertOptions(OrcSplitReader reader, long offset, long length) {
+		Reader.Options options = reader.getOptions();
+		Assert.assertEquals(offset, options.getOffset());
+		Assert.assertEquals(length, options.getLength());
+	}
+
+	private static OrcRowSplitReader newReaderWithStripes(
+			List<StripeInformation> stripes, FileInputSplit split) throws IOException {
+		TypeDescription td = TypeDescription.fromString(TEST_SCHEMA_FLAT);
+		return new OrcRowSplitReader(
+				new Configuration(),
+				td,
+				IntStream.range(0, td.getChildren().size()).toArray(),
+				new ArrayList<>(),
+				1024,
+				split.getPath(),
+				split.getStart(),
+				split.getLength()) {
+
+			@Override
+			List<StripeInformation> getStripes(Reader orcReader) {
+				return stripes;
+			}
+		};
 	}
 
 	@Test
 	public void testSplitStripesCustomSplits() throws IOException {
-		rowOrcInputFormat =
-			new OrcRowInputFormat(getPath(TEST_FILE_FLAT), TEST_SCHEMA_FLAT, new Configuration());
-
-		OrcRowInputFormat spy = spy(rowOrcInputFormat);
-
 		// mock list of stripes
 		List<StripeInformation> stripes = new ArrayList<>();
 		StripeInformation stripe1 = mock(StripeInformation.class);
@@ -256,25 +267,23 @@ public class OrcRowInputFormatTest {
 		stripes.add(stripe3);
 		stripes.add(stripe4);
 		stripes.add(stripe5);
-		doReturn(stripes).when(spy).getStripes(any());
 
-		// mock options to check configuration of ORC reader
-		Reader.Options options = spy(new Reader.Options());
-		doReturn(options).when(spy).getOptions(any());
-
-		spy.openInputFormat();
 		// split ranging 2 stripes
-		spy.open(new FileInputSplit(0, new Path(getPath(TEST_FILE_FLAT)), 0, 150, new String[]{}));
-		verify(options).range(eq(10L), eq(190L));
+		OrcRowSplitReader reader = newReaderWithStripes(
+				stripes, new FileInputSplit(0, new Path(getPath(TEST_FILE_FLAT)), 0, 150, new String[]{}));
+		assertOptions(reader, 10L, 190L);
 		// split ranging 0 stripes
-		spy.open(new FileInputSplit(1, new Path(getPath(TEST_FILE_FLAT)), 150, 10, new String[]{}));
-		verify(options).range(eq(0L), eq(0L));
+		reader = newReaderWithStripes(
+				stripes, new FileInputSplit(1, new Path(getPath(TEST_FILE_FLAT)), 150, 10, new String[]{}));
+		assertOptions(reader, 0L, 0L);
 		// split ranging 1 stripe
-		spy.open(new FileInputSplit(2, new Path(getPath(TEST_FILE_FLAT)), 160, 41, new String[]{}));
-		verify(options).range(eq(200L), eq(100L));
+		reader = newReaderWithStripes(
+				stripes, new FileInputSplit(2, new Path(getPath(TEST_FILE_FLAT)), 160, 41, new String[]{}));
+		assertOptions(reader, 200L, 100L);
 		// split ranging 2 stripe
-		spy.open(new FileInputSplit(3, new Path(getPath(TEST_FILE_FLAT)), 201, 299, new String[]{}));
-		verify(options).range(eq(300L), eq(200L));
+		reader = newReaderWithStripes(
+				stripes, new FileInputSplit(3, new Path(getPath(TEST_FILE_FLAT)), 201, 299, new String[]{}));
+		assertOptions(reader, 300L, 200L);
 	}
 
 	@Test
@@ -348,7 +357,7 @@ public class OrcRowInputFormatTest {
 
 		rowOrcInputFormat.selectFields(0, 4, 1);
 		rowOrcInputFormat.addPredicate(
-			new OrcRowInputFormat.Equals("_col1", PredicateLeaf.Type.STRING, "M"));
+			new OrcSplitReader.Equals("_col1", PredicateLeaf.Type.STRING, "M"));
 
 		byte[] bytes = InstantiationUtil.serializeObject(rowOrcInputFormat);
 		OrcRowInputFormat copy = InstantiationUtil.deserializeObject(bytes, getClass().getClassLoader());
@@ -376,45 +385,43 @@ public class OrcRowInputFormatTest {
 
 		// boolean pred
 		rowOrcInputFormat.addPredicate(
-			new OrcRowInputFormat.Equals("boolean1", PredicateLeaf.Type.BOOLEAN, false));
+			new OrcSplitReader.Equals("boolean1", PredicateLeaf.Type.BOOLEAN, false));
 		// boolean pred
 		rowOrcInputFormat.addPredicate(
-			new OrcRowInputFormat.LessThan("byte1", PredicateLeaf.Type.LONG, 1));
+			new OrcSplitReader.LessThan("byte1", PredicateLeaf.Type.LONG, 1));
 		// boolean pred
 		rowOrcInputFormat.addPredicate(
-			new OrcRowInputFormat.LessThanEquals("short1", PredicateLeaf.Type.LONG, 1024));
+			new OrcSplitReader.LessThanEquals("short1", PredicateLeaf.Type.LONG, 1024));
 		// boolean pred
 		rowOrcInputFormat.addPredicate(
-			new OrcRowInputFormat.Between("int1", PredicateLeaf.Type.LONG, -1, 65536));
+			new OrcSplitReader.Between("int1", PredicateLeaf.Type.LONG, -1, 65536));
 		// boolean pred
 		rowOrcInputFormat.addPredicate(
-			new OrcRowInputFormat.Equals("long1", PredicateLeaf.Type.LONG, 9223372036854775807L));
+			new OrcSplitReader.Equals("long1", PredicateLeaf.Type.LONG, 9223372036854775807L));
 		// boolean pred
 		rowOrcInputFormat.addPredicate(
-			new OrcRowInputFormat.Equals("float1", PredicateLeaf.Type.FLOAT, 1.0));
+			new OrcSplitReader.Equals("float1", PredicateLeaf.Type.FLOAT, 1.0));
 		// boolean pred
 		rowOrcInputFormat.addPredicate(
-			new OrcRowInputFormat.Equals("double1", PredicateLeaf.Type.FLOAT, -15.0));
+			new OrcSplitReader.Equals("double1", PredicateLeaf.Type.FLOAT, -15.0));
 		// boolean pred
 		rowOrcInputFormat.addPredicate(
-			new OrcRowInputFormat.IsNull("string1", PredicateLeaf.Type.STRING));
+			new OrcSplitReader.IsNull("string1", PredicateLeaf.Type.STRING));
 		// boolean pred
 		rowOrcInputFormat.addPredicate(
-			new OrcRowInputFormat.Equals("string1", PredicateLeaf.Type.STRING, "hello"));
+			new OrcSplitReader.Equals("string1", PredicateLeaf.Type.STRING, "hello"));
 
 		FileInputSplit[] splits = rowOrcInputFormat.createInputSplits(1);
 		rowOrcInputFormat.openInputFormat();
 
 		// mock options to check configuration of ORC reader
 		OrcRowInputFormat spy = spy(rowOrcInputFormat);
-		Reader.Options options = new Reader.Options();
-		doReturn(options).when(spy).getOptions(any());
 
 		spy.openInputFormat();
 		spy.open(splits[0]);
 
 		// verify predicate configuration
-		SearchArgument sarg = options.getSearchArgument();
+		SearchArgument sarg = spy.getReader().getOptions().getSearchArgument();
 		assertNotNull(sarg);
 		assertEquals("(and leaf-0 leaf-1 leaf-2 leaf-3 leaf-4 leaf-5 leaf-6 leaf-7 leaf-8)", sarg.getExpression().toString());
 		assertEquals(9, sarg.getLeaves().size());
@@ -437,11 +444,11 @@ public class OrcRowInputFormatTest {
 
 		rowOrcInputFormat.addPredicate(
 			// OR
-			new OrcRowInputFormat.Or(
+			new OrcSplitReader.Or(
 				// timestamp pred
-				new OrcRowInputFormat.Equals("time", PredicateLeaf.Type.TIMESTAMP, Timestamp.valueOf("1900-05-05 12:34:56.100")),
+				new OrcSplitReader.Equals("time", PredicateLeaf.Type.TIMESTAMP, Timestamp.valueOf("1900-05-05 12:34:56.100")),
 				// date pred
-				new OrcRowInputFormat.Equals("date", PredicateLeaf.Type.DATE, Date.valueOf("1900-12-25")))
+				new OrcSplitReader.Equals("date", PredicateLeaf.Type.DATE, Date.valueOf("1900-12-25")))
 			);
 
 		FileInputSplit[] splits = rowOrcInputFormat.createInputSplits(1);
@@ -449,14 +456,12 @@ public class OrcRowInputFormatTest {
 
 		// mock options to check configuration of ORC reader
 		OrcRowInputFormat spy = spy(rowOrcInputFormat);
-		Reader.Options options = new Reader.Options();
-		doReturn(options).when(spy).getOptions(any());
 
 		spy.openInputFormat();
 		spy.open(splits[0]);
 
 		// verify predicate configuration
-		SearchArgument sarg = options.getSearchArgument();
+		SearchArgument sarg = spy.getReader().getOptions().getSearchArgument();
 		assertNotNull(sarg);
 		assertEquals("(or leaf-0 leaf-1)", sarg.getExpression().toString());
 		assertEquals(2, sarg.getLeaves().size());
@@ -471,23 +476,21 @@ public class OrcRowInputFormatTest {
 			new OrcRowInputFormat(getPath(TEST_FILE_DECIMAL), TEST_SCHEMA_DECIMAL, new Configuration());
 
 		rowOrcInputFormat.addPredicate(
-			new OrcRowInputFormat.Not(
+			new OrcSplitReader.Not(
 				// decimal pred
-				new OrcRowInputFormat.Equals("_col0", PredicateLeaf.Type.DECIMAL, BigDecimal.valueOf(-1000.5))));
+				new OrcSplitReader.Equals("_col0", PredicateLeaf.Type.DECIMAL, BigDecimal.valueOf(-1000.5))));
 
 		FileInputSplit[] splits = rowOrcInputFormat.createInputSplits(1);
 		rowOrcInputFormat.openInputFormat();
 
 		// mock options to check configuration of ORC reader
 		OrcRowInputFormat spy = spy(rowOrcInputFormat);
-		Reader.Options options = new Reader.Options();
-		doReturn(options).when(spy).getOptions(any());
 
 		spy.openInputFormat();
 		spy.open(splits[0]);
 
 		// verify predicate configuration
-		SearchArgument sarg = options.getSearchArgument();
+		SearchArgument sarg = spy.getReader().getOptions().getSearchArgument();
 		assertNotNull(sarg);
 		assertEquals("(not leaf-0)", sarg.getExpression().toString());
 		assertEquals(1, sarg.getLeaves().size());
@@ -501,7 +504,7 @@ public class OrcRowInputFormatTest {
 			new OrcRowInputFormat(getPath(TEST_FILE_NESTED), TEST_SCHEMA_NESTED, new Configuration());
 
 		rowOrcInputFormat.addPredicate(
-			new OrcRowInputFormat.Equals("unknown", PredicateLeaf.Type.LONG, 42));
+			new OrcSplitReader.Equals("unknown", PredicateLeaf.Type.LONG, 42));
 	}
 
 	@Test
@@ -928,13 +931,13 @@ public class OrcRowInputFormatTest {
 
 		// read head and tail of file
 		rowOrcInputFormat.addPredicate(
-			new OrcRowInputFormat.Or(
-				new OrcRowInputFormat.LessThan("_col0", PredicateLeaf.Type.LONG, 10L),
-				new OrcRowInputFormat.Not(
-					new OrcRowInputFormat.LessThanEquals("_col0", PredicateLeaf.Type.LONG, 1920000L))
+			new OrcSplitReader.Or(
+				new OrcSplitReader.LessThan("_col0", PredicateLeaf.Type.LONG, 10L),
+				new OrcSplitReader.Not(
+					new OrcSplitReader.LessThanEquals("_col0", PredicateLeaf.Type.LONG, 1920000L))
 			));
 		rowOrcInputFormat.addPredicate(
-			new OrcRowInputFormat.Equals("_col1", PredicateLeaf.Type.STRING, "M"));
+			new OrcSplitReader.Equals("_col1", PredicateLeaf.Type.STRING, "M"));
 
 		FileInputSplit[] splits = rowOrcInputFormat.createInputSplits(1);
 		assertEquals(1, splits.length);
@@ -964,7 +967,7 @@ public class OrcRowInputFormatTest {
 		rowOrcInputFormat.selectFields(3, 0, 2);
 
 		rowOrcInputFormat.addPredicate(
-			new OrcRowInputFormat.LessThan("_col0", PredicateLeaf.Type.LONG, 10L));
+			new OrcSplitReader.LessThan("_col0", PredicateLeaf.Type.LONG, 10L));
 
 		FileInputSplit[] splits = rowOrcInputFormat.createInputSplits(1);
 		assertEquals(1, splits.length);
